Serve docs directly from live badge handler
