+++ endpoints/docs.ts
@@ -1,20 +1,26 @@
+import serveMarked from 'serve-marked'
 import serve404 from '../libs/serve-404'
-import serveHelp from '../libs/serve-help'
-import { liveBadgeList } from '../libs/badge-list'
-
-const badges = require('../static/.gen/badges.json')
+import genHelp from '../libs/gen-help'
 
 // Handles `/docs/:name`
 export default async function (req, res) {
   const [ , topic, name ] = req.url.split('/')
+  const helpMarkdown = genHelp(name)
 
-  if (liveBadgeList.includes(name)) {
+  if (helpMarkdown) {
     console.info(100, `${name}: ${req.url}`)
-    const foundBadge = badges.live.find(b => b.id === name)
-    if (foundBadge) {
-      return serveHelp(req, res, foundBadge)
-    }
+    return serveMarked(helpMarkdown, {
+      title: `${name} | Badgen`,
+      inlineCSS,
+    })(req, res)
   }
 
   serve404(req, res)
 }
+
+const inlineCSS = `
+  .markdown-body { max-width: 800px }
+  li > img { vertical-align: middle; margin: 0.2em 0; font-size: 12px }
+  li > img + a { font-family: monospace; font-size: 0.9em }
+  li > img + a + i { color: #AAA }
+`
