+++ libs/gen-help.ts
@@ -1,0 +1,35 @@
+import path from 'path'
+import { liveBadgeList } from './badge-list'
+
+const liveBadgeDB = liveBadgeList.reduce((accu, curr) => {
+  const { meta, handlers } = require(path.resolve(__dirname, `../endpoints/${curr}`))
+  const { title, examples, help } = meta
+  accu[curr] = {
+    id: curr,
+    title,
+    examples,
+    routes: Object.keys(handlers),
+    help
+  }
+  return accu
+}, {})
+
+export default function genHelp (id) {
+  const meta = liveBadgeDB[id]
+
+  if (!meta) {
+    return ''
+  }
+
+  const { examples, routes, help } = meta
+
+  const Docs = `# /${id}\n\n${help || ''}`
+  const schemeLinks = routes.map(r => `- \`${r}\``)
+  const Schemes = `## Schemes\n\n${schemeLinks.join('  \n')}`
+  // @ts-ignore
+  const exampleList = Object.entries(examples)
+    .map(([url, desc]) => `- ![${url}](${url}) [${url}](${url}) <i>${desc}</i>`)
+  const Examples = `## Examples\n\n${exampleList.join('\n')}`
+
+  return [Docs, Schemes, Examples].join('\n\n')
+}
